Implement the deposit items task and fix a lot of bugs

+++ src/main/java/com/willr27/blocklings/client/gui/control/controls/ItemControl.java
@@ -1,6 +1,7 @@
 package com.willr27.blocklings.client.gui.control.controls;
 
 import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.systems.RenderSystem;
 import com.willr27.blocklings.client.gui.control.Control;
 import com.willr27.blocklings.client.gui.util.GuiUtil;
 import com.willr27.blocklings.client.gui.util.ScissorStack;
@@ -46,7 +47,7 @@ public class ItemControl extends Control
         int x = (int) ((getPixelX() + getPixelWidth() / 2.0) / getGuiScale());
         int y = (int) ((getPixelY() + getPixelHeight() / 2.0) / getGuiScale());
 
-        double z = isDraggingOrAncestor() ? getDraggedControl().getDragZ() : getRenderZ();
+        float z = (float) (isDraggingOrAncestor() ? getDraggedControl().getDragZ() : getRenderZ());
 
         try
         {
